Convert an object to a simple feature or bounding box object

Both functions will pass a NULL value without returning an error.

Usage

as_sf(x, crs = NULL, ...)

as_bbox(x, crs = NULL, ...)

as_sfc(x, crs = NULL, ...)

Arguments

x

A sf, bbox, sfc, raster, sp, or dataframe object that can be converted into a simple feature or bounding box object. as_bbox() can also convert a vector with xmin, ymin, xmax, and ymax values.

crs

Coordinate reference system for sf or bbox obect to return

...

Additional parameters passed to sf::st_bbox() when calling as_bbox() or passed to sf::st_sf(), sf::st_as_sf(), or df_to_sf() for as_sf() (depending on class of x)