Get location of specified type

Filter by name or id or use a spatial filter based on an sf object or geocoded street address. Optionally you can use an index list to match the type to a named list of URLs or sf objects.

Usage

get_location(
  type,
  name = NULL,
  id = NULL,
  location = NULL,
  label = NULL,
  name_col = "name",
  id_col = "id",
  index = NULL,
  union = FALSE,
  ...
)

Arguments

type

Type of location to return. Type can be an sf object, e.g. a data frame with multiple neighborhoods or a character string that can be passed to get_location_data. If index is provided, character can also be a character string to match the name of a list.

name

Location name to return.

id

Location id to return. id is coerced to character or numeric to match the class of the id_col for type.

location

An address or sf or bbox object passed to st_filter. Any valid address or addresses are geocoded with geo), converted to a simple feature object, and then used as a spatial filter.

label

Label optionally added to "label" column; must be a length 1 or match the number of rows returned based on the other parameters. If union = TRUE, using label is recommended. Default: NULL

name_col

Column name in type with name values, Default: 'name' Required if name provided.

id_col

Column name in type with id values, Default: 'id'. Required if id is provided.

index

Optional list used to match type to data, Default: NULL

union

If TRUE, the location geometry is unioned with st_union and the names are combined into a single value. Default: FALSE.

...

Additional parameters passed to get_location_data if type is character and index is NULL.

Value

A simple feature object from data provided to type.

Examples

if (FALSE) {
if (interactive()) {
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
  # Check if type as sf object with name/id lookup works
  get_location(type = nc, name = "Warren", name_col = "NAME")
  get_location(type = nc, id = 37185, id_col = "FIPSNO")

  # Type also supports a range of other formats including

  # File path
  get_location(
    type = system.file("shape/nc.shp", package = "sf"),
    name = "Hertford",
    name_col = "NAME"
  )

  # Index name (if a named list of datasets, urls, or paths is passed to index)
  get_location(
    type = "smaller",
    name = "Hertford",
    name_col = "NAME",
    index = list(
      "smaller" = dplyr::filter(nc, AREA <= 0.10),
      "larger" = dplyr::filter(nc, AREA > 0.15)
    )
  )

  # url may require passing extra parameters to `get_location_data()`
  # In this example, no location information is passed to get_location() so it warns before returning all types
  get_location(
    type = "https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services/USA_Parks/FeatureServer/0",
    locationname_col = "NAME",
    locationname = "Chesapeake and Ohio Canal National Historic Park"
  )
}
}