Make a ggplot map using layer_location_data

Location is used as the data parameter of layer_location_data so this function is primarily appropriate for the layer_mapbox (geom = "mapbox").

Usage

make_location_map(
  location,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  paper = NULL,
  orientation = NULL,
  geom = "sf",
  basemap = TRUE,
  below = NULL,
  above = NULL,
  save = FALSE,
  name = NULL,
  label = NULL,
  prefix = NULL,
  postfix = NULL,
  filename = NULL,
  device = NULL,
  filetype = NULL,
  path = NULL,
  dpi = 300,
  ...
)

make_social_map(
  location,
  dist = NULL,
  diag_ratio = NULL,
  unit = NULL,
  asp = NULL,
  crs = 3857,
  image = NULL,
  platform = NULL,
  format = NULL,
  orientation = NULL,
  basemap = TRUE,
  geom = "mapbox",
  save = FALSE,
  name = NULL,
  filename = NULL,
  label = NULL,
  prefix = NULL,
  postfix = NULL,
  path = NULL,
  filetype = "jpeg",
  dpi = 72,
  ...
)

Arguments

location

sf object. If multiple areas are provided, they are unioned into a single sf object using sf::st_union

dist

buffer distance in units. Optional.

diag_ratio

ratio of diagonal distance of area's bounding box used as buffer distance. e.g. if the diagonal distance is 3000 meters and the "diag_ratio = 0.1" a 300 meter will be used. Ignored when dist is provided.

unit

Units for buffer. Supported options include "meter", "foot", "kilometer", and "mile", "nautical mile" Common abbreviations (e.g. "km" instead of "kilometer") are also supported. Distance in units is converted to units matching GDAL units for x; defaults to "meter"

asp

Aspect ratio of width to height as a numeric value (e.g. 0.33) or character (e.g. "1:3"). If numeric, get_asp() returns the same value without modification.

paper

Paper, Default: 'letter'.

orientation

Orientation "portrait", "landscape", or "square", Default: 'portrait'.

geom

ggplot2 geom to use, Default: 'sf'. Options include "sf" (geom_sf), "icon" (layer_icon / geom_sf_icon), "text" (geom_sf_text), "label" (geom_sf_label), "textsf", "labelsf", "text_repel", and "label_repel".

save

If TRUE, save file with ggsave_ext or ggsave_social, requires basemap = TRUE and filename is not NULL or ... include a name parameter. Default: FALSE

name

Plot name, used to create filename (if filename is NULL) using make_filename()

label

label for area (appended to data as a prefix if data is a string)

prefix

File name prefix. "date" adds a date prefix, "time" adds a date/time prefix; defaults to NULL.

postfix

File name postfix; defaults to NULL.

filename

File name to create on disk.

device

Device to use. Can either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).

filetype

file type supported by sf::read_sf. The file type must be provided for extdata and cached data.

path

Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.

dpi

Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.

...

Additional parameters passed to layer_location_data.

crs

Coordinate reference system of bounding box to return

image

Image size name, Default: NULL

platform

Social media platform, "Instagram", "Facebook", or "Twitter", Default: NULL

format

Image format, "post", "story", or "cover", Default: NULL