Read spatial data in a bounding box to a simple feature object

An extended version of read_sf that support reading spatial data based on a file path, url (for spatial data or ArcGIS FeatureServer or MapServer), or the data name and associated package. Optionally provide a bounding box to filter data.

Usage

read_sf_path(path, bbox = NULL, ...)

read_sf_url(url, bbox = NULL, ...)

read_sf_package(data, bbox = NULL, package, filetype = "gpkg", ...)

Arguments

path

file path

bbox

bbox object; Default: NULL. If bbox is provided, read_sf only returns features intersecting the bounding box.

...

additional parameters passed to read_sf

url

url for spatial data file or for ArcGIS FeatureServer or MapServer to access with get_esri_data()

data

character; name of data.

package

character; package name

filetype

file type supported by read_sf., Default: 'gpkg'. Required only if the data is in the cache directory or extdata system files.

Value

sf object

Details

This function looks for three types of package data: = Data loaded with the package

External data in the extdata system files folder.

Cached data in the cache directory returned by user_cache_dir