Measure, transform, and convert bounding boxes

Simple bounding box functions that you can use to:

Usage

sf_bbox_asp(bbox, orientation = FALSE)

sf_bbox_dist(bbox, from, to, units = FALSE)

sf_bbox_xdist(bbox, units = FALSE)

sf_bbox_ydist(bbox, units = FALSE)

sf_bbox_diagdist(bbox, units = FALSE)

sf_bbox_transform(bbox, crs)

sf_bbox_to_sf(bbox, sf_col = "geometry")

sf_bbox_to_sfc(bbox)

sf_bbox_to_wkt(bbox)

sf_bbox_to_lonlat_query(bbox, coords = c("longitude", "latitude"), crs = 4326)

sf_bbox_shift(
  bbox,
  x_nudge = 0,
  y_nudge = 0,
  side = c("all", "top", "bottom", "left", "right"),
  dir = NULL
)

sf_bbox_contract(bbox, x_nudge = 0, y_nudge = 0)

sf_bbox_expand(bbox, x_nudge = 0, y_nudge = 0)

Arguments

bbox

bbox object

orientation

If TRUE, sf_bbox_asp returns a suggested orientation based on aspect ratio (< 0.9 "portrait"; > 1.1 "landscape"; else "square"); defaults to FALSE.

from, to

xy pairs (e.g. c("xmax", "ymax) defining points to measure distance from and to.

units

If TRUE, distance functions return with units. If FALSE (default), distance functions return numeric values.

crs

coordinate reference system to use for query; default 4326

coords

query column names with coordinates. e,g, c("X", "Y") or c("lat", "lon")

x_nudge, y_nudge

Length 1 or 2 numeric vector; unitless.

side

one or more sides to shift: "top", "bottom", "left", "right", or "all"

dir

If "in", contract the bbox by x_nudge and y_nudge. If "out", expand the bbox by x_nudge and y_nudge. If dir is not NULL; absolute values are used for x_nudge and y_nudge. Defaults to NULL.

Value

sf object

Details

Measure distances (sf_bbox_dist, sf_bbox_xdist, sf_bbox_ydist, and sf_bbox_diagdist)

Get an aspect ratio (sf_bbox_asp)

Transform the coordinate reference system (sf_bbox_transform)

Convert a bounding box to a SQL style query (sf_bbox_to_lonlat_query), well known text (sf_bbox_to_wkt), or a simple feature object (sf_bbox_to_sf)

Shift, expand, or contract a bounding box (sf_bbox_shift, sf_bbox_expand, sf_bbox_contract)

The only functions with additional parameters are sf_bbox_to_lonlat_query and sf_bbox_dist. All other functions only take a bbox parameter.

See also