Convert simple feature object to data frame with coordinates or data frame to simple feature

Helper function to convert a simple feature object to data frame by dropping geometry, converting geometry to well known text, or (if the geometry type is not POINT) getting coordinates for a centroid or point on surface. If an sfc object is provided, sf_to_sfc() provides coordinates but the "drop" geometry option is not supported.

Usage

sf_to_df(
  x,
  crs = 4326,
  coords = c("lon", "lat"),
  geometry = "centroid",
  keep_all = TRUE
)

df_to_sf(x, crs = 4326, coords = c("lon", "lat"), into = NULL, sep = ",")

Arguments

x

A sf or sfc object or a data frame with lat/lon coordinates in a single column or two separated columns

crs

cCordinate reference system, Default: 4326

coords

Coordinate columns for input dataframe or output sf object (if geometry is 'centroid' or 'point') Default: c("lon", "lat")

geometry

Type of geometry to include in data frame. options include "drop", "wkt", "centroid", "point", Default: 'centroid'

keep_all

If FALSE, drop all columns other than those named in coords, Default: TRUE

into

If coords is a single column name with both longitude and latitude, into is used as the names of the new columns that coords is separated into. Passed to separate

sep

If coords is a single column name with both longitude and latitude, sep is used as the separator between coordinate values. Passed to separate

Value

sf_to_df() returns a data frame with geometry dropped or converted to wkt or coordinates for the centroid or point on surface; df_to_sf()returns a simple feature object with POINT geometry

See also

Examples

if (FALSE) {
if (interactive()) {
  nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
  nc_df <- ggspatial::df_spatial(nc)

  df_to_sf(nc_df, coords = c("x", "y"))

  nc_df$xy <- paste(nc_df$x, nc_df$y, sep = ",")

  df_to_sf(nc_df, coords = "xy", into = c("lon", "lat"))
}
}