Save a ggplot and update file EXIF metadata

Save a plot or map then update the EXIF metadata for the title, author, and create data. ggsave_ext() also supports creating a file name based on a sentence case name with spaces (e.g. "Baltimore city map") and appending a label (e.g. "baltcity") as a prefix to the output file name.

Usage

ggsave_ext(
  name = NULL,
  label = NULL,
  plot = last_plot(),
  filename = NULL,
  path = NULL,
  device = "png",
  paper = NULL,
  orientation = "portrait",
  scale = 1,
  width,
  height,
  units = "in",
  dpi = 300,
  bg = NULL,
  exif = FALSE,
  title = NULL,
  author,
  args = NULL,
  prefix = "date",
  ...
)

Arguments

name

name of plot, used to create filename (if filename is not provided)

label

label is appended to filename as a prefix; defaults to NULL

plot

Plot to save, defaults to last plot displayed.

filename

File name to create on disk.

path

Path of the directory to save plot to: path and filename are combined to create the fully qualified file name. Defaults to the working directory.

device

Device to use. Can either be a device function (e.g. png), or one of "eps", "ps", "tex" (pictex), "pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).

paper

paper matching name from paper_sizes (e.g. "letter"). Not case sensitive

orientation

"portrait", "landscape", or "square"

scale

Multiplicative scaling factor.

width

Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.

height

Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.

units

Plot size in units ("in", "cm", "mm", or "px"). If not supplied, uses the size of current graphics device.

dpi

Plot resolution. Also accepts a string input: "retina" (320), "print" (300), or "screen" (72). Applies only to raster output types.

bg

Background colour. If NULL, uses the plot.background fill value from the plot theme.

exif

If TRUE, edit exif metadata for exported file using the exifr package, Default: FALSE

title

title of plot or map, added to exif metadata, Default: NULL

author

author of plot or map, added to exif metadata,

args

args passed to exiftoolr, If args is not NULL, title and author are ignored, Default: NULL

prefix

If "date", the date from Sys.Date() is appended to filename as a prefix (before label). If "time", Sys.time() is appended.

...

additional parameters passed to ggsave

See also

Examples

if (FALSE) {
if (interactive()) {
  neighborhoods <-
    get_location(
      type = "neighborhoods",
      package = "mapbaltimore"
    )

  ggplot2::ggplot() +
    ggplot2::geom_sf(data = neighborhoods)

  ggsave_ext(
    name = "neighborhoods",
    label = "Baltimore City",
    author = "Thomas H. Poppleton",
    device = "pdf",
    paper = "letter"
  )
}
}