Read spatial data in a bounding box to a simple feature object

An extended version of sf::read_sf() that support reading spatial data based on a file path, URL, or the data name and associated package. Optionally provide a bounding box to filter data (not supported for all data types). If a file or path is provided for a GeoJSON file, the read_sf_geojson function (using the suggested geojsonsf package) is used.

Usage

read_sf_ext(..., bbox = NULL)

read_sf_pkg(data, bbox = NULL, package = NULL, filetype = "gpkg", ...)

read_sf_path(path, bbox = NULL, ...)

read_sf_excel(
  path,
  sheet = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  geo = FALSE,
  address = "address",
  ...
)

read_sf_csv(
  path,
  url = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  geo = FALSE,
  address = "address",
  ...
)

read_sf_url(url, bbox = NULL, coords = NULL, ...)

read_sf_geojson(url, path = NULL, geojson = NULL, bbox = NULL, ...)

read_sf_gist(url, id = NULL, bbox = NULL, ...)

read_sf_download(
  url,
  filename,
  bbox = NULL,
  path = NULL,
  filetype = "geojson",
  prefix = "date",
  method = "auto",
  unzip = FALSE,
  ...
)

read_sf_gsheet(
  url,
  sheet = NULL,
  ss = NULL,
  bbox = NULL,
  coords = c("lon", "lat"),
  ask = FALSE,
  geo = FALSE,
  address = "address",
  ...
)

Arguments

...

additional parameters passed to multiple functions; see details.

bbox

A bounding box object; Default: NULL. If "bbox" is provided, read_sf only returns features intersecting the bounding box.

data

character; name of data; used by read_sf_pkg() only

package

character; package name; used by read_sf_pkg() only

filetype

file type supported by sf::read_sf()., Default: 'gpkg'; used by read_sf_pkg() only and required only if the data is in the package cache directory or extdata system files.

path

A file path; used by read_sf_path() only.

sheet

Sheet to read. Either a string (the name of a sheet), or an integer (the position of the sheet). Ignored if the sheet is specified via range. If neither argument specifies the sheet, defaults to the first sheet.

coords

Character vector with coordinate values; used for read_sf_url() if the "url" is a Google Sheet.

geo

If TRUE, use address_to_sf to geocode address column; defaults to FALSE.

address

Address column name, Default: 'address'

url

A url for a spatial data file, a GitHub gist, or a ArcGIS FeatureServer or MapServer to access with get_esri_data()

geojson

string or vector of GeoJSON, or a URL or file pointing to a geojson file

id

(character) A gist id, or a gist URL

filename

File name; if file name is NULL, name and file type are both required.

prefix

File name prefix. "date" adds a date prefix, "time" adds a date/time prefix; defaults to NULL.

method

Method to be used for downloading files. Current download methods are "internal", "wininet" (Windows only) "libcurl", "wget" and "curl", and there is a value "auto": see ‘Details’ and ‘Note’.

The method can also be set through the option "download.file.method": see options().

unzip

If TRUE, url must be a zip file that is downloaded, unzipped into a temporary directory (created with tempdir()), and then read to a file using the specified file type.

ss

Something that identifies a Google Sheet:

its file id as a string or drive_id

a URL from which we can recover the id

a one-row dribble, which is how googledrive represents Drive files

an instance of googlesheets4_spreadsheet, which is what gs4_get() returns

Processed through as_sheets_id().

ask

If TRUE, ask for the name of the Google Sheet to read if ss is not provided to read_sf_gsheet.

Details

Reading data from a url:

read_sf_url supports multiple types of urls:

A MapServer or FeatureServer URL

A URL for a GitHub gist with a single spatial data file

A URL for a spatial data file or a CSV file

A Google Sheets URL

Reading data from a package:

read_sf_pkg looks for three types of package data:

Data loaded with the package

External data in the extdata system files folder.

Cached data in the cache directory returned by rappdirs::user_cache_dir

Additional ... parameters:

read_sf_pkg and read_sf_download both pass additional parameters to read_sf_path which supports query, name_col, name, and table. name and name_col are ignored if a query parameter is provided. If table is not provided, a expected layer name is created based on the file path.

read_sf_url pass the where, name_col, and name for any ArcGIS FeatureServer or MapServer url (passed to get_esri_data) or sheet if the url is for a Google Sheet (passed to googlesheets4::read_sheet), or a query or wkt filter parameter if the url is some other type (passed to sf::read_sf).

read_sf_ext is a flexible function that only has bbox as a named parameters and all other parameters in ... are passed to one of the other overedge read_sf functions.

See also