Convert between simple feature and data frame objects

Helper function to convert a simple feature object to data frame by dropping geometry, converting geometry to well known text, or (if the geometry type is not POINT) getting coordinates for a centroid or point on surface. If an sfc object is provided,the "drop" geometry option is not supported. check_coords is a helper function used by df_to_sf to suggest the appropriate coordinate column names based on the column names in the provided data frame.

Usage

sf_to_df(
  x,
  crs = 4326,
  coords = c("lon", "lat"),
  geometry = "centroid",
  keep_all = TRUE
)

df_to_sf(
  x,
  crs = NULL,
  coords = c("lon", "lat"),
  df_crs = 4326,
  into = NULL,
  sep = ",",
  rev = TRUE,
  remove_coords = FALSE
)

check_coords(x = NULL, coords = NULL, rev = FALSE)

Arguments

x

A sf or sfc object or a data frame with lat/lon coordinates in a single column or two separated columns.

crs

Cordinate reference system to return, Default: 4326 for sf_to_df and NULL for df_to_sf

coords

Coordinate columns for input dataframe or output sf object (if geometry is 'centroid' or 'point') Default: c("lon", "lat").

geometry

Type of geometry to include in data frame. options include "drop", "wkt", "centroid", "point", Default: 'centroid'.

keep_all

If FALSE, drop all columns other than those named in coords, Default: TRUE.

df_crs

For df_to_sf, coordinate reference system used by coordinates or well known text in data frame.

into

If coords is a single column name with both longitude and latitude, into is used as the names of the new columns that coords is separated into. Passed to tidyr::separate.

sep

If coords is a single column name with both longitude and latitude, sep is used as the separator between coordinate values. Passed to tidyr::separate.

rev

If TRUE, reverse c("lat", "lon") coords to c("lon", "lat"). check_coords only.

remove_coords

For df_to_sf, if TRUE, remove the coordinate columns after converting a data frame to simple feature object; defaults to FALSE.

Value

sf_to_df() returns a data frame with geometry dropped or converted to wkt or coordinates for the centroid or point on surface; df_to_sf()returns a simple feature object with POINT geometry.

See also

Examples

nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"))
#> Reading layer `nc' from data source 
#>   `/Library/Frameworks/R.framework/Versions/4.1-arm64/Resources/library/sf/shape/nc.shp' 
#>   using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27

nc_coords <-
  suppressWarnings(sf::st_coordinates(sf::st_centroid(nc)))

nc_df <-
  dplyr::bind_cols(sf::st_drop_geometry(nc), nc_coords)

nc_df <- janitor::clean_names(nc_df)
df_to_sf(nc_df, coords = c("x", "y"), remove_coords = TRUE)
#> Simple feature collection with 100 features and 14 fields
#> Attribute-geometry relationship: 14 constant, 0 aggregate, 0 identity
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -84.05986 ymin: 34.07671 xmax: -75.8095 ymax: 36.49111
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>     area perimeter cnty cnty_id        name  fips fipsno cress_id bir74 sid74
#> 1  0.114     1.442 1825    1825        Ashe 37009  37009        5  1091     1
#> 2  0.061     1.231 1827    1827   Alleghany 37005  37005        3   487     0
#> 3  0.143     1.630 1828    1828       Surry 37171  37171       86  3188     5
#> 4  0.070     2.968 1831    1831   Currituck 37053  37053       27   508     1
#> 5  0.153     2.206 1832    1832 Northampton 37131  37131       66  1421     9
#> 6  0.097     1.670 1833    1833    Hertford 37091  37091       46  1452     7
#> 7  0.062     1.547 1834    1834      Camden 37029  37029       15   286     0
#> 8  0.091     1.284 1835    1835       Gates 37073  37073       37   420     0
#> 9  0.118     1.421 1836    1836      Warren 37185  37185       93   968     4
#> 10 0.124     1.428 1837    1837      Stokes 37169  37169       85  1612     1
#>    nwbir74 bir79 sid79 nwbir79                   geometry
#> 1       10  1364     0      19  POINT (-81.49823 36.4314)
#> 2       10   542     3      12 POINT (-81.12513 36.49111)
#> 3      208  3616     6     260 POINT (-80.68573 36.41252)
#> 4      123   830     2     145 POINT (-76.02719 36.40714)
#> 5     1066  1606     3    1197 POINT (-77.41046 36.42236)
#> 6      954  1838     5    1237 POINT (-76.99472 36.36142)
#> 7      115   350     2     139 POINT (-76.23402 36.40122)
#> 8      254   594     2     371 POINT (-76.70446 36.44428)
#> 9      748  1190     2     844 POINT (-78.11042 36.39693)
#> 10     160  2038     5     176 POINT (-80.23429 36.40042)
nc_df$xy <- paste(nc_df$x, nc_df$y, sep = ",")
df_to_sf(nc_df, coords = "xy", into = c("lon", "lat"))
#> Simple feature collection with 100 features and 18 fields
#> Attribute-geometry relationship: 18 constant, 0 aggregate, 0 identity
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -84.05986 ymin: 34.07671 xmax: -75.8095 ymax: 36.49111
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>     area perimeter cnty cnty_id        name  fips fipsno cress_id bir74 sid74
#> 1  0.114     1.442 1825    1825        Ashe 37009  37009        5  1091     1
#> 2  0.061     1.231 1827    1827   Alleghany 37005  37005        3   487     0
#> 3  0.143     1.630 1828    1828       Surry 37171  37171       86  3188     5
#> 4  0.070     2.968 1831    1831   Currituck 37053  37053       27   508     1
#> 5  0.153     2.206 1832    1832 Northampton 37131  37131       66  1421     9
#> 6  0.097     1.670 1833    1833    Hertford 37091  37091       46  1452     7
#> 7  0.062     1.547 1834    1834      Camden 37029  37029       15   286     0
#> 8  0.091     1.284 1835    1835       Gates 37073  37073       37   420     0
#> 9  0.118     1.421 1836    1836      Warren 37185  37185       93   968     4
#> 10 0.124     1.428 1837    1837      Stokes 37169  37169       85  1612     1
#>    nwbir74 bir79 sid79 nwbir79         x        y       lon      lat
#> 1       10  1364     0      19 -81.49823 36.43140 -81.49823 36.43140
#> 2       10   542     3      12 -81.12513 36.49111 -81.12513 36.49111
#> 3      208  3616     6     260 -80.68573 36.41252 -80.68573 36.41252
#> 4      123   830     2     145 -76.02719 36.40714 -76.02719 36.40714
#> 5     1066  1606     3    1197 -77.41046 36.42236 -77.41046 36.42236
#> 6      954  1838     5    1237 -76.99472 36.36142 -76.99472 36.36142
#> 7      115   350     2     139 -76.23402 36.40122 -76.23402 36.40122
#> 8      254   594     2     371 -76.70446 36.44428 -76.70446 36.44428
#> 9      748  1190     2     844 -78.11042 36.39693 -78.11042 36.39693
#> 10     160  2038     5     176 -80.23429 36.40042 -80.23429 36.40042
#>                      geometry
#> 1   POINT (-81.49823 36.4314)
#> 2  POINT (-81.12513 36.49111)
#> 3  POINT (-80.68573 36.41252)
#> 4  POINT (-76.02719 36.40714)
#> 5  POINT (-77.41046 36.42236)
#> 6  POINT (-76.99472 36.36142)
#> 7  POINT (-76.23402 36.40122)
#> 8  POINT (-76.70446 36.44428)
#> 9  POINT (-78.11042 36.39693)
#> 10 POINT (-80.23429 36.40042)